Dz4

+++ app/src/main/java/com/happydroid/happytodo/data/repository/TodoItemsRepository.kt
@@ -1,21 +1,31 @@
 package com.happydroid.happytodo.data.repository
 
 import android.app.Application
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
 import android.util.Log
+import androidx.work.Constraints
+import androidx.work.ExistingWorkPolicy
+import androidx.work.NetworkType
+import androidx.work.OneTimeWorkRequest
+import androidx.work.WorkManager
 import com.happydroid.happytodo.data.datasource.FakeDataSource
 import com.happydroid.happytodo.data.local.LocalStorage
-import com.happydroid.happytodo.data.local.TodoItemDao
 import com.happydroid.happytodo.data.model.ErrorCode
+import com.happydroid.happytodo.data.model.Mapper
 import com.happydroid.happytodo.data.model.TodoItem
-import com.happydroid.happytodo.data.model.TodoResult
+import com.happydroid.happytodo.data.model.TodoItemsResult
 import com.happydroid.happytodo.data.model.toTodoElementRequestNW
-import com.happydroid.happytodo.data.model.toTodoItemNW
+import com.happydroid.happytodo.data.model.toTodoItemNetwork
 import com.happydroid.happytodo.data.network.TodoApiFactory
-import com.happydroid.happytodo.data.network.model.ResponseNW
+import com.happydroid.happytodo.data.network.model.ResponseNetwork
 import com.happydroid.happytodo.data.network.model.RevisionHolder
-import com.happydroid.happytodo.data.network.model.TodoListRequestNW
-import com.happydroid.happytodo.data.network.model.TodoListResponseNW
-import com.happydroid.happytodo.data.network.model.toTodoResult
+import com.happydroid.happytodo.data.network.model.TodoListRequestNetwork
+import com.happydroid.happytodo.data.network.model.TodoListResponseNetwork
+import com.happydroid.happytodo.data.network.model.toTodoItemsResult
+import com.happydroid.happytodo.data.workers.SyncWorker
+import com.happydroid.happytodo.di.AppScope
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
@@ -25,18 +35,27 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import retrofit2.Response
-import java.net.InetAddress
 import java.net.UnknownHostException
 import java.util.Date
+import javax.inject.Inject
 
-class TodoItemsRepository private constructor(application: Application){
 
-    private val fakeDataSource : FakeDataSource= FakeDataSource()
-    private val todoItemDao: TodoItemDao = LocalStorage.getDatabase(application).todoItems()
-    private val apiRemote = TodoApiFactory.retrofitService
-    private val _todoItemsResult = MutableStateFlow(TodoResult())
-    val todoItemsResult: StateFlow<TodoResult> = _todoItemsResult
-    var attempt = 0
+private const val DELEAY_NOTIFICATION = 5000L
+
+/**
+ * This class is responsible for managing the persistence and retrieval of todo items.
+ */
+@Suppress("UNCHECKED_CAST")
+@AppScope
+class TodoItemsRepository @Inject constructor(application: Application, private val fakeDataSource: FakeDataSource, localStorage: LocalStorage, todoApiFactory : TodoApiFactory) {
+
+    private val todoItemDao = localStorage.todoItems()
+    private val apiRemote = todoApiFactory.retrofitService
+    private val workManager = WorkManager.getInstance(application)
+    private val connectivityManager = application.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+    private val _todoItemsResult = MutableStateFlow(TodoItemsResult())
+    val todoItemsResult: StateFlow<TodoItemsResult> = _todoItemsResult
+    private var attempt = 0
 
     init {
         CoroutineScope(Dispatchers.IO).launch {
@@ -51,71 +70,97 @@ class TodoItemsRepository private constructor(application: Application){
             }
             saveAllTodoItemsToRemote()
 
-            if (!isInternetAvailable()) {
+            if (!isOnline()) {
                 Log.i("HappyTodo", "No internet connection")
-                delay(5000L)
+                delay(DELEAY_NOTIFICATION)
                 val message = ErrorCode.NO_CONNECTION
                 _todoItemsResult.value = todoItemsResult.value.copy(errorMessages = listOf(message))
             }
         }
     }
 
-    suspend fun observeLocalTodoItems() {
+    private suspend fun observeLocalTodoItems() {
         todoItemDao.observeAll().collect { todoItems ->
-            _todoItemsResult.value = TodoResult(data = todoItems)
+            _todoItemsResult.value = TodoItemsResult(data = todoItems)
         }
     }
 
+    private suspend fun syncOnConnection() {
+        val message = ErrorCode.NO_CONNECTION
+        val oldResult = todoItemsResult.value
+        val newErrorMessages = todoItemsResult.value.errorMessages.toMutableList().apply {
+            add(message)
+        }
 
-    suspend fun saveAllTodoItemsToRemote(){
+        _todoItemsResult.value = oldResult.copy(errorMessages = newErrorMessages)
+
+
+        val myWorkRequest = OneTimeWorkRequest.Builder(SyncWorker::class.java)
+            .setConstraints(
+                Constraints.Builder()
+                    .setRequiredNetworkType(NetworkType.CONNECTED)
+                    .build()
+            )
+            .build()
+        workManager.enqueueUniqueWork("HappySyncWorker", ExistingWorkPolicy.REPLACE, myWorkRequest)
+    }
+
+    suspend fun saveAllTodoItemsToRemote() {
         try {
-            val response = apiRemote.updateAll(
-                TodoListRequestNW(todoItemsResult.value.data.map { it.toTodoItemNW()  }))
-            handleResponse(response as Response<ResponseNW>)
+            if (isOnline()){
+                val response = apiRemote.updateAll(
+                    TodoListRequestNetwork(todoItemsResult.value.data.map { it.toTodoItemNetwork() })
+                )
+                handleResponse(response as Response<ResponseNetwork>)
+            }else{
+                syncOnConnection()
+            }
+        }catch (e: UnknownHostException){
+            syncOnConnection()
         } catch (e: Exception) {
             Log.e("TodoItemsRepository.saveAllTodoItemsToRemote()", "Exception: ${e.message}")
         }
     }
 
-    suspend fun handleResponse(response: Response<ResponseNW>) {
+    private suspend fun handleResponse(response: Response<ResponseNetwork>) {
         if (response.isSuccessful) {
             attempt = 0
-            val responseNW: ResponseNW? = response.body()
-            responseNW?.revision?.let { RevisionHolder.revision = it }
+            val responseNetwork: ResponseNetwork? = response.body()
+            responseNetwork?.revision?.let { RevisionHolder.revision = it }
 
         } else {
-            val errorCode = response.code()
+            val errorCode = Mapper().mapToErrorCode(response.code())
             Log.i("HappyTodo", "Код ошибки сервера: " + errorCode)
 
             attempt++
-            var message : ErrorCode? = null
+            var message: ErrorCode? = null
 
             withContext(Dispatchers.IO) {
                 delay(1000L * attempt * attempt)
 
-                if(errorCode == 400 || errorCode == 500){
+                if (errorCode == ErrorCode.ERROR_401) {
                     try {
                         apiRemote.fetchAll() // запрос, чтобы просто получить Revision
                         saveAllTodoItemsToRemote()   // в идеале, бэк должен смержить все данные
                         fetchFromRemoteApi() //получаем обновленные данные с сервера
-                    }catch (e: Exception) {
+                    } catch (e: Exception) {
                         Log.e("TodoItemsRepository", "Exception: ${e.message}")
                     }
 
-                }else if(errorCode == 401){
-                    message = ErrorCode.ERROR_401
-                }else if(errorCode == 404){
+                } else if (errorCode == ErrorCode.ERROR_500) {
+                    message = ErrorCode.ERROR_500
+                } else if (errorCode == ErrorCode.ERROR_404) {
                     message = ErrorCode.ERROR_404
-                }else if (!isInternetAvailable()){
+                } else if (!isOnline()) {
                     message = ErrorCode.NO_CONNECTION
-                }else{
-
+                    syncOnConnection()
+                } else {
                     message = ErrorCode.UNKNOW_ERROR
                 }
             }
             val oldResult = todoItemsResult.value
             val newErrorMessages = todoItemsResult.value.errorMessages.toMutableList().apply {
-                message?.let{add(it)} // Добавляем новый элемент в список
+                message?.let { add(it) } // Добавляем новый элемент в список
             }
             _todoItemsResult.value = oldResult.copy(errorMessages = newErrorMessages)
 
@@ -123,35 +168,46 @@ class TodoItemsRepository private constructor(application: Application){
         }
     }
 
-    suspend fun loadFakesTodoItems() {
-         withContext(Dispatchers.IO) {
-             val loadedList = fakeDataSource.loadTodoItems()
-             _todoItemsResult.value = TodoResult(loadedList, listOf(ErrorCode.LOAD_FROM_HARDCODED_DATASOURCE))
+    private suspend fun loadFakesTodoItems() {
+        withContext(Dispatchers.IO) {
+            val loadedList = fakeDataSource.loadTodoItems()
+            _todoItemsResult.value =
+                TodoItemsResult(loadedList, listOf(ErrorCode.LOAD_FROM_HARDCODED_DATASOURCE))
         }
     }
 
     suspend fun fetchFromRemoteApi() {
         withContext(Dispatchers.IO) {
             try {
-                val response = apiRemote.fetchAll()
-                handleResponse(response as Response<ResponseNW>)
+                if(isOnline()){
 
-                if(response.isSuccessful){
-                    val todoListResponseNW: TodoListResponseNW? = response.body()
-                    val newErrorMessages = todoItemsResult.value.errorMessages.toMutableList().apply {
-                        add(ErrorCode.LOAD_FROM_REMOTE)
+                    val response = apiRemote.fetchAll()
+                    handleResponse(response as Response<ResponseNetwork>)
+
+                    if (response.isSuccessful) {
+                        val todoListResponseNW: TodoListResponseNetwork? = response.body()
+                        val newErrorMessages =
+                            todoItemsResult.value.errorMessages.toMutableList().apply {
+                                add(ErrorCode.LOAD_FROM_REMOTE)
+                            }
+
+                        // Обновляем данные
+                        _todoItemsResult.value =
+                            todoListResponseNW?.toTodoItemsResult(newErrorMessages) ?: TodoItemsResult(
+                                emptyList(), listOf(ErrorCode.UNKNOW_ERROR)
+                            )
+                        todoItemDao.addAll(todoItemsResult.value.data)
+
+                    } else {
+                        // компилятор не дает удалить этот блок else
+                        // обработка в handleResponse()
                     }
-
-                    // Обновляем данные
-                    _todoItemsResult.value = todoListResponseNW?.toTodoResult(newErrorMessages) ?: TodoResult(
-                        emptyList(), listOf(ErrorCode.UNKNOW_ERROR)
-                    )
-
                 }else{
-                    // компилятор не дает удалить этот блок else
-                    // обработка в handleResponse()
+                    syncOnConnection()
                 }
 
+            }catch (e: UnknownHostException){
+                syncOnConnection()
             } catch (e: Exception) {
                 Log.e("TodoItemsRepository.fetchFromRemote()", "Exception: ${e.message}")
             }
@@ -159,10 +215,10 @@ class TodoItemsRepository private constructor(application: Application){
 
     }
 
-    fun removeMessageFromQueue(messageId : ErrorCode){
+    fun removeMessageFromQueue(messageId: ErrorCode) {
         _todoItemsResult.update { todoItemsResult ->
             val errorMessages = todoItemsResult.errorMessages
-                .filterNot { it.stringResId == messageId.stringResId}
+                .filterNot { it.stringResId == messageId.stringResId }
             todoItemsResult.copy(errorMessages = errorMessages)
         }
     }
@@ -174,26 +230,26 @@ class TodoItemsRepository private constructor(application: Application){
         _todoItemsResult.value = _todoItemsResult.value.copy(data = newItems)
 
         withContext(Dispatchers.IO) {
-            try{
+            try {
                 todoItemDao.deleteById(idTodoItem)
-
-                val response = apiRemote.deleteItem (idTodoItem)
-                handleResponse(response as Response<ResponseNW>)
+                if (isOnline()) {
+                    val response = apiRemote.deleteItem(idTodoItem)
+                    handleResponse(response as Response<ResponseNetwork>)
+                }else{
+                    syncOnConnection()
+                }
+            }catch (e: UnknownHostException){
+                syncOnConnection()
             } catch (e: Exception) {
                 Log.e("TodoItemsRepository.deleteTodoItem()", "Exception: ${e.message}")
             }
         }
     }
 
-
     suspend fun getTodoItem(idTodoItem: String): TodoItem? {
         return withContext(Dispatchers.Default) {
             todoItemsResult.value.data.find { it.id == idTodoItem }
         }
-    }
-
-    suspend fun getTodoItems(): List<TodoItem> {
-        return withContext(Dispatchers.Default) { todoItemsResult.value.data }
     }
 
     suspend fun changeStatusTodoItem(idTodoItem: String, isDone: Boolean) {
@@ -203,99 +259,86 @@ class TodoItemsRepository private constructor(application: Application){
                 else todoItem
             }
         }
-        _todoItemsResult.value = TodoResult(newItems)
+        _todoItemsResult.value = TodoItemsResult(newItems)
 
-
-        try{
+        try {
             withContext(Dispatchers.IO) {
                 val newItem = getTodoItem(idTodoItem)
-                 newItem?.let{
-                    todoItemDao.editTodoItem(newItem)
-                    val response = apiRemote.updateItem (idTodoItem, it.toTodoElementRequestNW())
-                    handleResponse(response as Response<ResponseNW>)
+                newItem?.let {
+                    todoItemDao.updateTodoItem(newItem)
+                    if (isOnline()) {
+                        val response = apiRemote.updateItem(idTodoItem, it.toTodoElementRequestNW())
+                        handleResponse(response as Response<ResponseNetwork>)
+                    }else{
+                        syncOnConnection()
+                    }
                 }
             }
-        }catch (e: Exception) {
+
+        }catch (e: UnknownHostException){
+                syncOnConnection()
+        } catch (e: Exception) {
             Log.e("TodoItemsRepository.changeStatusTodoItem()", "Exception: ${e.message}")
         }
     }
 
-    suspend fun addOrUpdateTodoItem(todoItem: TodoItem) {
-        var isExisted = false
-        var response : Response<ResponseNW>
-
-        val newItems = withContext(Dispatchers.Default) {
-
-            val currentList = todoItemsResult.value.data.toMutableList()
-            val index = currentList.indexOfFirst { it.id == todoItem.id }
-            if (index != -1) {
-                // Элемент найден, перезаписываем его
-                currentList[index] = todoItem
-                isExisted = true
-            } else {
-                currentList.add(todoItem)
-            }
-            currentList.toList()
-        }
-        _todoItemsResult.value = TodoResult(newItems)
-        try{
+    suspend fun addTodoItem(todoItem: TodoItem) {
+        try {
             withContext(Dispatchers.IO) {
-                if (isExisted){
-                    todoItemDao.editTodoItem(todoItem)
-                    response = apiRemote.updateItem(todoItem.id, todoItem.toTodoElementRequestNW()) as Response<ResponseNW>
+                todoItemDao.addTodoItem(todoItem)
+                if (isOnline()) {
+                    val response =
+                        apiRemote.addItem(todoItem.toTodoElementRequestNW()) as Response<ResponseNetwork>
+                    handleResponse(response)
                 }else{
-                    todoItemDao.addTodoItem(todoItem)
-                    response = apiRemote.addItem(todoItem.toTodoElementRequestNW()) as Response<ResponseNW>
+                    syncOnConnection()
                 }
-                handleResponse(response)
             }
-        }catch (e: Exception) {
+        }
+        catch (e: UnknownHostException){
+            syncOnConnection()
+        }
+        catch (e: Exception) {
             Log.e("TodoItemsRepository.addOrUpdateTodoItem()", "Exception: ${e.message}")
         }
     }
 
-    suspend fun updateTodoItem(newTodoItem: TodoItem)  {
-        val newItems = withContext(Dispatchers.Default) {
-            todoItemsResult.value.data.map { todoItem ->
-                if (todoItem.id == newTodoItem.id) newTodoItem
-                else todoItem
-            }
-        }
-        _todoItemsResult.value = TodoResult(newItems)
-        try{
+    suspend fun updateTodoItem(todoItem: TodoItem) {
+        try {
             withContext(Dispatchers.IO) {
-                val newItem = getTodoItem(newTodoItem.id)
-                newItem?.let{
-                    todoItemDao.editTodoItem(newItem)
-                    val response = apiRemote.updateItem (newItem.id, it.toTodoElementRequestNW())
-                    handleResponse(response as Response<ResponseNW>)
+                todoItemDao.updateTodoItem(todoItem)
+                if (isOnline()) {
+                    val response =
+                        apiRemote.updateItem(todoItem.id, todoItem.toTodoElementRequestNW()) as Response<ResponseNetwork>
+                    handleResponse(response)
+                }else{
+                    syncOnConnection()
                 }
             }
-        }catch (e: Exception) {
-            Log.e("TodoItemsRepository.updateTodoItem()", "Exception: ${e.message}")
         }
-    }
-
-    private fun isInternetAvailable(): Boolean {
-        return try {
-            val address = InetAddress.getByName("ya.ru")
-            !address.equals("")
-        } catch (e: UnknownHostException) {
-            Log.e("TodoItemsRepository.isInternetAvailable()", "Exception: ${e.message}")
-            false
+        catch (e: UnknownHostException){
+            syncOnConnection()
+        }
+        catch (e: Exception) {
+            Log.e("TodoItemsRepository.addOrUpdateTodoItem()", "Exception: ${e.message}")
         }
     }
 
 
-
-    companion object {
-        @Volatile
-        private var instance: TodoItemsRepository? = null
-
-        fun getInstance(application: Application): TodoItemsRepository {
-            return instance ?: synchronized(this) {
-                instance ?: TodoItemsRepository(application).also { instance = it }
-            }
-        }
+    private fun isOnline(): Boolean {
+        val capabilities = connectivityManager.getNetworkCapabilities(connectivityManager.activeNetwork)
+        return capabilities?.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET) == true
     }
+
+
+//    companion object {
+//        @Volatile
+//        private var instance: TodoItemsRepository? = null
+//        @Inject
+//        fun getInstance(application: Application): TodoItemsRepository {
+//            return instance ?: synchronized(this) {
+//                instance ?: TodoItemsRepository(application).also { instance = it }
+//            }
+//        }
+//    }
 }
